third_party/WebKit/LayoutTests/fullscreen/api/element-request-fullscreen-and-remove.html

<!DOCTYPE html>
<title>Element#requestFullscreen() followed by removing the element</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<div id="target"></div>
<script>
// Adapted from https://github.com/w3c/web-platform-tests/pull/4250
// TODO(foolip): Remove this test when the above is imported and passing.
async_test(t => {
  const target = document.getElementById("target");

  document.onfullscreenchange = t.step_func_done(() => {
    assert_equals(document.fullscreenElement, null);
  });
  document.onfullscreenerror = t.unreached_func("fullscreenchange event");

  trusted_click(t.step_func(() => {
    target.requestFullscreen();
    target.remove();
  }), document.body);
});
</script>
